Enhancment of coco evaluation scale_boxes function for yolox tutorial
